// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/AgeRangeDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AgeRangeDto = {
    ageRangeId: number;
    rangeName: string;
    rangeLeft: string;
    rangeRight: string;
    modified: string;
    modifiedByUsername?: string | null;
};